#585 Move tests related to the Literal PSyIR node to tests/psyir/nodes/literal_test.py
